changed to REAL

+++ arbo_lib/db/schema.sql
@@ -1,9 +1,9 @@
 CREATE TABLE task_models (
     task_name VARCHAR(255) PRIMARY KEY,
-    t_base_1 FLOAT,            -- Baseline time at s=1 (seconds)
-    p_obs FLOAT,               -- Current dynamic parallelizable portion
-    c_startup FLOAT,           -- Fixed overhead (e.g., pod spin-up)
-    alpha FLOAT DEFAULT 0.3,   -- Learning rate for p_obs
+    t_base_1 REAL,            -- Baseline time at s=1 (seconds)
+    p_obs REAL,               -- Current dynamic parallelizable portion
+    c_startup REAL,           -- Fixed overhead (e.g., pod spin-up)
+    alpha REAL DEFAULT 0.3,   -- Learning rate for p_obs
     last_updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     base_input_quantity FLOAT DEFAULT 1,
     sample_count INT DEFAULT 0
@@ -13,12 +13,12 @@ CREATE TABLE execution_history (
     id SERIAL PRIMARY KEY,
     task_name VARCHAR(255) REFERENCES task_models(task_name),
     parallelism INT NOT NULL ,           -- s
-    input_scale_factor FLOAT NOT NULL,  -- gamma
-    cluster_load INT NOT NULL,          -- L_cluster
-    total_duration FLOAT NOT NULL,      -- Actual T (Wall Clock Time)
-    residual FLOAT NOT NULL,            -- T_actual - T_amdahl
-    cost_metric FLOAT,                 -- Cost of run
-    p_snapshot FLOAT,                  -- p of execution
+    input_scale_factor REAL NOT NULL,  -- gamma
+    cluster_load REAL NOT NULL,          -- L_cluster
+    total_duration REAL NOT NULL,      -- Actual T (Wall Clock Time)
+    residual REAL NOT NULL,            -- T_actual - T_amdahl
+    cost_metric REAL,                 -- Cost of run
+    p_snapshot REAL,                  -- p of execution
     recorded_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
